feat: custom kaleidoscope config for Atreus
A config that makes heavy use of QuKeys to optimize home row usage and maximize
the number of keys on the first layer. Might need to tweak the QuKeys
configuration in case it turns our chords and rolls result in mistyped
characters but this should serve as a nice start.