chore(hyprland): configure for nvidia
Hyprland no longer depends on a patched version of the nvidia driver so
we can safely use the default package and configure it accordingly.

I had initially set things up so that Hyprland runs on the Intel GPU so
that I can enable / disable the kernel modules for the Nvidia GPU
dynamically and attach it to a Window VM but I've not yet managed to get
that working. Possibly because I'm not using Nvidia Prime since the
Hyprland wiki says that it doesn't play nicely with the Hyprland
compositor.

To be continued..