feat(fonts): properly configure standard font names
For better system wide configuration and less tweaking of individual
applications the convention is to define which fonts the standard serif,
sans-serif and monospace aliases should fall back to. This seems to have
simplified things somewhat and also resolved an issue with the web
versionf Todoist rendering fonts with weird spacing issues.